Move createSubScope to its own file
